week2,3 work (freq; cis eqtl)
we mainly tried to do some analysis on the dataset to be familiar with the dataset and analysis method
we used plink and python to get the alleles frequency. also we do cleaning on the dataset and run the linear regression(cis eqtl analysis). this is for the locuszoom graph